One module to rule them all
