--- src/Interface/keys.jl
@@ -1,100 +1,0 @@
-
-"""
-    step_key(x)
-
-Returns the step size of the keys of `x`.
-
-## Examples
-```jldoctest
-julia> using AxisIndices
-
-julia> AxisIndices.step_key(Axis(1:2:10))
-2
-
-julia> AxisIndices.step_key(rand(2))
-1
-
-julia> AxisIndices.step_key([1])  # LinearIndices are treate like unit ranges
-1
-```
-"""
-@inline step_key(x::AbstractVector) = _step_keys(keys(x))
-_step_keys(ks) = step(ks)
-_step_keys(ks::LinearIndices) = 1
-
-"""
-    axes_keys(x) -> Tuple
-
-Returns the keys corresponding to all axes of `x`.
-
-## Examples
-```jldoctest
-julia> using AxisIndices
-
-julia> axes_keys(AxisArray(ones(2,2), (2:3, 3:4)))
-(2:3, 3:4)
-
-julia> axes_keys(Axis(1:2))
-(1:2,)
-```
-"""
-axes_keys(x) = map(keys, axes(x))
-axes_keys(x::AbstractUnitRange) = (keys(x),)
-
-"""
-    axes_keys(x, i)
-
-Returns the axis keys corresponding of ith dimension of `x`.
-
-## Examples
-```jldoctest
-julia> using AxisIndices
-
-julia> axes_keys(AxisArray(ones(2,2), (2:3, 3:4)), 1)
-2:3
-```
-"""
-axes_keys(x, i) = axes_keys(x)[i]
-@inline axes_keys(x, i::Symbol) =  axes_keys(x, dims(x, i)) # TODO test for types stability
-
-
-"""
-    keys_type(x, i)
-
-Retrieves axis keys of the ith dimension of `x`.
-
-## Examples
-```jldoctest
-julia> using AxisIndices
-
-julia> keys_type(AxisArray([1], ["a"]), 1)
-Array{String,1}
-```
-"""
-keys_type(::T, i) where {T} = keys_type(T, i)
-keys_type(::Type{T}, i) where {T} = keys_type(axes_type(T, i))
-
-"""
-    keys_type(x)
-
-Retrieves the type of the keys of `x`.
-
-## Examples
-```jldoctest
-julia> using AxisIndices
-
-julia> keys_type(Axis(1:2))
-UnitRange{Int64}
-
-julia> keys_type(typeof(Axis(1:2)))
-UnitRange{Int64}
-
-julia> keys_type(UnitRange{Int})
-Base.OneTo{Int64}
-```
-"""
-keys_type(::T) where {T} = keys_type(T)
-keys_type(::Type{T}) where {T<:AbstractRange} = OneTo{Int}  # default for things is usually LinearIndices{1}
-@inline function keys_type(::Type{T}) where {T<:AbstractArray}
-    return Tuple{ntuple(i -> keys_type(T, i), Val(ndims(T)))...}
-end
